refactor(Lotto): 메서드 반환형 유의미한 값으로 변경

+++ src/main/java/lotto/model/Lotto.java
@@ -33,8 +33,8 @@ public class Lotto {
 
     }
 
-    public long matchCount(Lotto lotto){
-        return numbers.stream()
+    public int matchCount(Lotto lotto){
+        return (int) numbers.stream()
                 .filter(lotto::hasNumber)
                 .count();
     }
